refactor(parts): move yaml utils out of lifecycle
